A pretty hefty commit, with a bunch of stuff.
Some styling, especially with lists of items and when there is focus on an item. Added a browser focus, which is an absolutely positioned UI with a drop shadow that hovers over any item in order to provide the sense of focusing on it.

Split out the views of ItemView into ItemValueView and ItemReferenceView, which brought about the ItemReference, a proxy for an Item before the item has actually loaded. There is still the chance that getReferencedItem returns null, which is a problem. Rather the ItemReference should proxy for the item methods, with subscribe, getProperty etc, and then when the id of the item being references has finally loaded it will hand over all subscriptions to the real item.

The lists of items, such as the drawer and the label list, now use a ListComponent, which keeps track of all the items and handles keystrokes to manage focus properly.

Focusing on an item is now done using browser.itemFocus, as mentioned above