Expose sets through the fin API with three methods: observeSet, addToSet and removeFromSet.
Since queries were already using sets as the underlying data structure, this patch adds support for set mutations and repurpose them for queries.

In order to simplify the unique identification of item properties and query IDs, we now use the key as the channel name for redis subscriptions, rather than having a channel name for e.g. a query that is different from its key.

This patch succesfully abstracts some of the complexities of the state and state change flows